requirements.txt
scrapy
requests
beautifulsoup4
pyyaml
psycopg2